+++ src/hooks/reservations/calculateEndTime.ts
@@ -15,7 +15,7 @@ export const calculateEndTime =
 
     const service = await req.payload.findByID({
       id: serviceId,
-      collection: config.slugs.services as 'reservation-services',
+      collection: config.slugs.services as 'services',
       req,
     })